+++ apps/web/src/components/tx-flow-2/flows/index.ts
@@ -1,4 +1,3 @@
 import dynamic from 'next/dynamic'
 
-export const ChangeThresholdFlow = dynamic(() => import('./ChangeThreshold'))
 export const ConfirmTxFlow = dynamic(() => import('./ConfirmTx'))
