--- apps/web/src/components/tx-flow/flows/ChangeThreshold/context.tsx
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const ChangeThresholdReviewContext = createContext<{ newThreshold?: number }>({})
