+++ apps/web/src/components/tx-flow/flows/ChangeThreshold/index.tsx
@@ -1,13 +1,8 @@
-import TxLayout from '@/components/tx-flow/common/TxLayout'
-import type { TxStep } from '@/components/tx-flow/common/TxLayout'
-import ReviewChangeThreshold from '@/components/tx-flow/flows/ChangeThreshold/ReviewChangeThreshold'
-import useTxStepper from '@/components/tx-flow/useTxStepper'
 import SaveAddressIcon from '@/public/images/common/save-address.svg'
 import useSafeInfo from '@/hooks/useSafeInfo'
-import { ChooseThreshold } from '@/components/tx-flow/flows/ChangeThreshold/ChooseThreshold'
-import { ConfirmTxDetails } from '@/components/tx/ConfirmTxDetails'
-import { useMemo } from 'react'
-import { TxFlowType } from '@/services/analytics'
+import { createDefaultTxFlow, type SubmitCallbackWithData } from '@/components/tx-flow-2/createTxFlow'
+import { ChooseThreshold } from './ChooseThreshold'
+import { SETTINGS_EVENTS, trackEvent } from '@/services/analytics'
 
 export enum ChangeThresholdFlowFieldNames {
   threshold = 'threshold',
@@ -17,44 +12,32 @@ export type ChangeThresholdFlowProps = {
   [ChangeThresholdFlowFieldNames.threshold]: number
 }
 
+const TxFlow = createDefaultTxFlow<ChangeThresholdFlowProps>()
+
 const ChangeThresholdFlow = () => {
-  const { safe } = useSafeInfo()
+  const {
+    safe: { threshold, owners },
+  } = useSafeInfo()
 
-  const { data, step, nextStep, prevStep } = useTxStepper<ChangeThresholdFlowProps>(
-    {
-      [ChangeThresholdFlowFieldNames.threshold]: safe.threshold,
-    },
-    TxFlowType.CHANGE_THRESHOLD,
-  )
+  const trackEvents = (newThreshold: number) => {
+    trackEvent({ ...SETTINGS_EVENTS.SETUP.OWNERS, label: owners.length })
+    trackEvent({ ...SETTINGS_EVENTS.SETUP.THRESHOLD, label: newThreshold })
+  }
 
-  const steps = useMemo<TxStep[]>(
-    () => [
-      {
-        txLayoutProps: { title: 'New transaction' },
-        content: <ChooseThreshold key={0} params={data} onSubmit={(formData) => nextStep(formData)} />,
-      },
-      {
-        txLayoutProps: { title: 'Confirm transaction' },
-        content: <ReviewChangeThreshold key={1} params={data} onSubmit={() => nextStep(data)} />,
-      },
-      {
-        txLayoutProps: { title: 'Confirm transaction details', fixedNonce: true },
-        content: <ConfirmTxDetails key={2} onSubmit={() => {}} showMethodCall />,
-      },
-    ],
-    [nextStep, data],
-  )
+  const handleSubmit: SubmitCallbackWithData<ChangeThresholdFlowProps> = ({ data }) => {
+    trackEvents(data?.threshold || threshold)
+  }
 
   return (
-    <TxLayout
-      subtitle="Change threshold"
+    <TxFlow
+      initialData={{ threshold }}
       icon={SaveAddressIcon}
-      step={step}
-      onBack={prevStep}
-      {...(steps?.[step]?.txLayoutProps || {})}
+      subtitle="Change threshold"
+      onSubmit={handleSubmit}
+      showMethodCall
     >
-      {steps.map(({ content }) => content)}
-    </TxLayout>
+      <ChooseThreshold key={0} />
+    </TxFlow>
   )
 }
